* (bug 30921) Fix breakage in dumpHTML on Special:Categories
SkinTemplate::buildContentNavigationUrls() uses $request->getRequestURL() when building the special tab on special page; so when building Special:Categories list we ended up with a fatal exception, as FauxRequest doesn't support that function.
Setting up the article context with a modified DumpFauxRequest that's specific to the article -- and does know to return the URL -- seems to resolve this.